Move configuration code to test
The code is to be distributed as a library, and any run-time
configurations should be kept off the main code.
