+++ src/test/java/uk/gov/hmcts/reform/api/deprecated/Application.java
@@ -1,4 +1,4 @@
-package hello;
+package uk.gov.hmcts.reform.api.deprecated;
 
 import java.util.Arrays;
